계획 수정

+++ script/lib/Motor/Motor.cpp
@@ -20,19 +20,14 @@ control::control(int pwm_pin, int motor_pin_a, int motor_pin_b, int servo_pin )
     pinMode(motor_pin_B_, OUTPUT);
     pinMode(pwm_pin_, OUTPUT); 
 
-    // pwm.pinFreq1(pwm_pin);
-    // pwm.setFreq1(PWM_DC);
-    
-    // pwm.pinDuty(pwm_pin_, 0);
     steering_servo.attach(servo_pin_);
 
 }
 
-void control::run(int pwm_duty)
+void control::run(double pwm_duty)
 {   
     
     pwm_duty_ = pwm_duty;
-    // pwm_duty_ = myMap(pwm_duty_, 0, 4000);
     if(pwm_duty > 0)
     {  
         digitalWrite(motor_pin_A_, HIGH);
@@ -44,7 +39,6 @@ void control::run(int pwm_duty)
         digitalWrite(motor_pin_B_, HIGH);
     }
     analogWrite(pwm_pin_, abs(pwm_duty_));
-    // pwm.pinDuty(pwm_pin_, abs(pwm_duty_));
 }
 
 float control::steer(float steering_angle)
