+++ test_due/lib/Motor/Motor.h
@@ -3,7 +3,7 @@
 class control
 {
 private:
-    int pwm_duty_;
+    double pwm_duty_;
     int pwm_pin_;
     int motor_pin_A_;
     int motor_pin_B_;
@@ -15,7 +15,7 @@ private:
 
 public:
     control(int pwm_pin, int motor_pin_a, int motor_pin_b, int servo_pin);
-    void run(int pwm);
+    void run(double pwm);
     float steer(float steering_angle);
     float mapFloat(float x, float in_min, float in_max, float out_min, float out_max);
 };
